Fixing Makefile and test/main.cc
Motivation:
Currently test/main.cc includes the test programs that it runs which
makes it impossible to set a breakpoint in the bugger.

Modifications:
Removed the includes from test/main.cc and instead compile main.cc with
the test source in question.

Result:
It's now possible to set a breakpoint in lldb.