Adding ctags settings to include Node.js tags file
Motivation:
Currently there is no way to use ctags from within the sources of this
project which is means that it is not possible to jump to code in
Node.js or V8.

Modifications:
Added node/tags ctags file to .vimrc

Result:
Navigating using CTRL+[ allows for jumping to Node.js/V8 sources.